+++ lab5_q1/lab5_q1/Celebrity.h
@@ -9,6 +9,6 @@ public:
 	Celebrity();
 	Celebrity(string name, int favorite_food, int allergic_food);
 	string show_name();
-	int buy_food(int food, int quantity);
+	float buy_food(int food, int quantity);
 };
