+++ lab5_q1/lab5_q1/Customer.h
@@ -19,6 +19,6 @@ public:
 	Customer();
 	Customer(string name);
 	virtual string show_name();
-	virtual int buy_food(int food, int quantity);
+	virtual float buy_food(int food, int quantity);
 };
